// src/awards/enums/award.enum.ts
export enum AwardType {
    MOTIVADO = 'motivado',
    MAESTRO = 'maestro',
    LEITOR = 'leitor',
    ESCRITOR = 'escritor',
    CONFIANTE = 'confiante',
    CORAJOSO = 'corajoso',
    SENSIVEL = 'sensivel',
    FANTASTICO = 'fantastico',
    LUTADOR = 'lutador',
    SUPER = 'super',
    INCRIVEL = 'incrivel',
    EXCELENTE = 'excelente',
    BRILHANTE = 'brilhante',
    GIGANTE = 'gigante',
    COMPETENTE = 'competente',
    BOMBASTICO = 'bombastico',
    SENSACIONAL = 'sensacional',
    PERFEITO = 'perfeito',
}